Add a new mixin for handling multiple targets
